[None][chore] Move FinishReasonsHandler into finish_reasons.py
The handler was an 847-line class nested inside TorchSampler that never
touched the outer instance: it takes plain sizes at construction and owns
all of its own state (the finish-reason store, the stop-word rule and
past-token buffers, their resize/refresh logic, and the end-id /
max-length / stop-word device checks).

Promote it to a module-level class in its own submodule, alongside
token_ban.py and top_p_decay.py. Only the construction site and the two
self-referencing type annotations change; the test monkeypatch targets
move from TorchSampler.FinishReasonsHandler to the imported class.
Formatting reflows follow from the reduced indentation. No behavior
change.
